feat: code refactor in gluestack-ui and gluestack-cli packages
