+++ packages/gluestack-cli/package.json
@@ -1,0 +1,70 @@
+{
+  "name": "gluestack-cli",
+  "keywords": [
+    "react",
+    "react-native",
+    "expo",
+    "storybook",
+    "nextjs",
+    "cross-platform",
+    "cli"
+  ],
+  "description": "A CLI tool for easily adding components from gluestack to your projects.",
+  "version": "0.0.1",
+  "license": "MIT",
+  "main": "dist/index.js",
+  "files": [
+    "dist",
+    "src"
+  ],
+  "engines": {
+    "node": ">=18"
+  },
+  "scripts": {
+    "start": "tsc --watch",
+    "watch": "tsc --watch",
+    "build": "tsc",
+    "prepare": "tsc",
+    "test": "echo 'Test passed'"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "tsdx lint"
+    }
+  },
+  "bin": {
+    "gluestack-cli": "dist/index.js"
+  },
+  "prettier": {
+    "printWidth": 80,
+    "semi": true,
+    "singleQuote": true,
+    "trailingComma": "es5"
+  },
+  "author": "gluestack",
+  "dependencies": {
+    "@antfu/ni": "^0.21.12",
+    "@clack/prompts": "^0.6.3",
+    "commander": "^12.0.0",
+    "find-package-json": "^1.2.0",
+    "fs-extra": "^11.1.0",
+    "simple-git": "^3.16.0"
+  },
+  "devDependencies": {
+    "@size-limit/preset-small-lib": "^8.2.4",
+    "@types/find-package-json": "^1.2.3",
+    "@types/fs-extra": "^11.0.1",
+    "@types/node": "^18.16.16",
+    "husky": "^8.0.3",
+    "jest": "^29.5.0",
+    "jest-junit": "^15.0.0",
+    "react-native": "^0.70.7",
+    "react-native-web": "^0.18.12",
+    "size-limit": "^8.2.4",
+    "supertest": "^6.3.3",
+    "tsdx": "^0.14.1",
+    "tslib": "^2.5.0",
+    "typescript": "^5.0.4",
+    "util": "^0.12.5"
+  }
+}
