--- packages/gluestack-ui/src/util/get-package-info.ts
@@ -1,9 +1,0 @@
-import path from 'path';
-import fs from 'fs-extra';
-import { type PackageJson } from 'type-fest';
-
-export function getPackageInfo() {
-  const packageJsonPath = path.join('package.json');
-
-  return fs.readJSONSync(packageJsonPath) as PackageJson;
-}
