--- packages/gluestack-ui/src/util/get-package-manager.ts
@@ -1,13 +1,0 @@
-import { detect } from '@antfu/ni';
-
-export async function getPackageManager(
-  targetDir: string
-): Promise<'yarn' | 'pnpm' | 'bun' | 'npm'> {
-  const packageManager = await detect({ programmatic: true, cwd: targetDir });
-
-  if (packageManager === 'yarn@berry') return 'yarn';
-  if (packageManager === 'pnpm@6') return 'pnpm';
-  if (packageManager === 'bun') return 'bun';
-
-  return packageManager ?? 'npm';
-}
